Add ability to specify and launch an HPC job via PBS
This commit adds a new class to the HPC module, PBS. The PBS class is derived
from the HPC class. The HPC base class is to serve as an interface for
executing HPC commands on the command line (i.e. qsub and in the future,
sbatch), ensuring jobs are canceled on SIGINT, and logging commands it preforms
as well as STDOUT and STDERR of the qsub.

The PBS are responsible for translating the arguments used in HPC launch_job
into runnable PBS script (e.g. script.pbs) and calling HPC to run it.

Future commits could implement other HPC types in the similar way as the PBS
class.