Update test class to parse .cfg files in the test directory
This commit breaks all tests that do not use the Test class as a base class, so
all tests will need to be updated to be able to be loaded and ran.

However, this commit updates the Test class to automatically parse files within
the test directory of a loaded test. Files that have the .cfg extension will be
parsed.

Configuration files will need be in the format as described here:
https://docs.python.org/3/library/configparser.html#module-configparser