Add module with MPAS utilities
This commit adds a new module, `smarts.utils.mpas`, which contains functions and
attributes that are commonly used when running MPAS in a test. For instance,
this module has a function to: check the output of a log file, edit an MPAS
streams file. It also contains attributes which contain the
`datetime.datetime.strftime`, `datetime.datetime.strptime` format strings for
history, restart and diagnostic files.