Fix format of run directory
Before this commit, the run directory had a dash '-' between the date and time
of the run directory, which made it hard to distinguish the two. This commit
changes the connecting dash to an underscore '_' and places a period before the
date. This is similar to the default MPAS datetime formats.