Allow ability to skip version check when loading env
Currently, when an environment is loaded, Smarts will check to see if the
compiler is loaded successfully by invoking the version for that command and
trying to check the version specified in the env file and the version in the
output.

It's probably not a necessary check to do, but also it causes issues because
currently on my MacBook pro my gcc and gfortran report different version. GCC
reports a clang version and GFrotran reports a gnu version.

Thus this change allows the loader to skip such checks.