Allow global, non-modset variables to be set in environment file
This commit allows an environment file the ability to specify environment
variables that are not associated with a Modset.

There might be a few uses for this, but specifically this change is to help me
be able to test on a cray machine, where all the compilers are aliased as
`ftn`.